feat(ocr): move OCR settings to dedicated screen under Learning section
