+++ i18n/src/commonMain/moko-resources/base/strings.xml
@@ -411,7 +411,12 @@
     <string name="pref_dict_korean_parser_legacy">Legacy parser</string>
     <string name="pref_dict_korean_parser_analyzer">Morpheme analyzer</string>
     <string name="pref_category_ocr">OCR</string>
-    <string name="pref_ocr_summary">Auto-OCR, cache management</string>
+    <string name="pref_ocr_summary">Engine, appearance, cache</string>
+    <string name="pref_category_engine">Engine</string>
+    <string name="pref_category_ocr_boxes">OCR Boxes</string>
+    <string name="pref_category_screen_overlay">Screen Overlay</string>
+    <string name="pref_category_behavior">Behavior</string>
+    <string name="pref_category_cache_storage">Cache Storage</string>
     <string name="pref_ocr_auto_on_download">Auto-OCR on download</string>
     <string name="pref_ocr_auto_on_download_summary">Automatically OCR chapters when downloading</string>
     <string name="pref_dict_ocr_box_scale">OCR box scale</string>
